test: Set lower success interval for the Watch Trigger Test
